Envs expanded to store prototypes
On the path to ptr / int interpolability.

In addition to the expansion a few other changes have been made:

EnvAST has been moved and is now scope, the former for type def order.
Basic fns now store arg info, and the two existing print fns have been updated.
To handle this pk fns for types are needed outside a driver, so these have been duplicated in the type header --- to be unified soon.