Casts
A short recap:

The book makes no distinction between pointers and ints (and also uses 32 bit ints, though that's not at issue here).
This is an issue with LLVM codegen as there is (as implemented) as distinction between pointers and ints.
This must be resolved somehow, as the (eventual) goal is to compile microC as in the book.

The solution settled on is to make casts are required, and specifically as part of the AST.
Previous commits set up the foundations for casts, and these are now made for fn calls.

So, this commit:

- Introduces a cast node, and everything required for the node to be concrete (i.e. codegen).
  The node stores an expression and a type, with the interpretation of the node being a cast of the expr to the type.
  The requirements of the cast are determined by inspecting the type of the expr.

- Adds functionality for casts to be automatically made.
  Specifically, casts are made when adding a call to the AST, and whether a cast is required is done by checking the type of the args given against the prototype.
  Any mismatch then calls a new `pk_ExprCast` fn which throws on an unsupported cast.

- Revises the `access` fn a little for clarity.
  Specifically, the redundant fn has been removed and the fresh fn now uses a switch for exhaustivity.

This all seems to work ok, though some additional work is required for casts in LLVM as (e.g.) comparisons are i1 while microC expects an int (/ i64).