Introduction of boolean type, and type header cleanup
LLVM codegen represents booleans as single bit ints.
microC (following the book) expects i64 (or really i32) ints.

The simple approach here seems to be expanding the way ints and pointers were handled.
That is, by introducing casts as required throughout the AST.

This commit adds a boolen type, and tidies the type header, though does not integrate the boolean type into AST gen in any way.

The header tidying is mostly cosmetic.
Though, an important refinement is using `llvm` within `defaultgen` to ensure `llvm` is the single source for how to codegen a type.