Bool casting basics
Boolean ops now (mostly, if not allly) using boolean types in the AST (and also in codegen) with appropriate casts made for fn calls, at least.

This is mostly in line with the way pointers are handled.

Type resolution for expressions now (mostly) directly returns a type, to help fix types.
And, some type related flow has been rewritten with switches to offload some reasoning onto the compiler.

ex2.c now compiles and runs ok, with expected results for the comparisons.

Though, some additional work on accessing / resolution is required, as in some places casts are made where derefs should be...