Eval true / false & defaultgen adjustments
defaultgen is now a method for all expressions and uses LLVM's `getNullValue` fn.
In turn, eval true evaluates whether an expression is not equal to its null value, while eval false evals whether the expression is equal.

In addition, eval true / false now perform access.
This requires some reworking of access, as eval true / false is a method on an expression and can't pass a shared pointer to itself.
So, for now, and maybe forever, access now takes a raw poiner, with most uses now updated to get the pointer.
Messy, but it works mostly ok.