Access refinments
The `access` method now returns type together with value.
And, a companion `access_type` method is implemented.
Ideally these would be combined, but for now they must be manually kept in sync.

Elsewhere the realisation that access on an address dereferences the address is not the semantics I'm after took a little reflection.