Dereference fixes (maybe)
The difficulty, generally, has been avoiding the issue of taking the address of an object.
This requires careful consideration about how the access method works, and likewise when it makes sense to take the address of and dereference an object.

From a slightly broader perspective, the issue is the absence of self-reference for objects.
Where, most of the complexity is due to avoiding self reference.

So, address of and dereference now anticipate and avoid instances of self reference.

Other issues were absence of pointer decay and incorrect types for pointer arithmetic.

Things seem mostly functional, though can certainly be improved.

In hindsight, too, it was a mistake to merge address and deref into prim1.
From a parsing perspective this makes some sense.
Though, from a functional perspective, the special treatment of these ops means it'd be nice to keep them conceptually separate from 'regular' ops.