parser fixes
My guess is that my pass at the resolution fn from the book is quite different.
As, types such as `int *var[i]` failed to parse correctly.
And, expressions such as `*(int var[10])` would parse.

Lack of testing, abundance of despair, well...

This certainly complicated the issues with resolving addressing and dereferencing.

;_; (sincerely, other things have been rough and while progress is being made so much of this is a mess...)