dereference & index fixes
deref loads, addr avoids a load.

with the previous fixes to the parser, etc. there's less confusion.
of note, no special case is required for &* / *& and so it has been removed.

the semantics really are to straightforwardly access or a deref and do nothing on an addr.
and, in the case of access, to do exactly this, with no change.

the index fix is to handle indicies exactly as pointer offsets.
the code for these should be unified at some point.