Simplified access(_type) methods
Primarily by merging cases.
Returns have also been unified, and a piece of dead code removed.
For clarity checks have been removed, as it's highly unlikely they would be effictive.
In particular, access is intended to be called from a expression, and could be recast as a method on expressions...