Fix shadowing issues
The initial approach to declarations was to clear the env before a declaration.
A benefit of this is that codegen for existing vars can be skipped.
This remained even though the approach did not, and use of shadowed vars was instead use of the original var.
This is now fixed, with an additional comment added.

ex6 now passes tests, and tests for ex5 has been revised (as it turns out I didn't read the source carefully)...